test cases passed

+++ auth/test/setup.js
@@ -4,19 +4,36 @@ const { MongoMemoryServer } = require("mongodb-memory-server");
 let mongo;
 
 beforeAll(async () => {
+  // ⏱ increase timeout for Mongo startup
+  jest.setTimeout(20000);
+  
+
   mongo = await MongoMemoryServer.create();
   const uri = mongo.getUri();
-  await mongoose.connect(uri);
+  process.env.MONGO_URI = uri;
+  process.env.JWT_SECRET = "test_jwt_secret"
+
+  await mongoose.connect(uri, {
+    dbName: "jest",
+  });
 });
 
 afterEach(async () => {
+  if (mongoose.connection.readyState !== 1) return;
+
   const collections = await mongoose.connection.db.collections();
-  for (let collection of collections) {
+
+  for (const collection of collections) {
     await collection.deleteMany({});
   }
 });
 
 afterAll(async () => {
-  await mongoose.connection.close();
-  await mongo.stop();
+  if (mongoose.connection.readyState === 1) {
+    await mongoose.connection.close();
+  }
+
+  if (mongo) {
+    await mongo.stop();
+  }
 });
